Added hidden Current Wounds attribute for Tokens
Added an attribute, CurrentWounds, and the sheet worker to keep its value updated, for people who prefer to track their wounds with a single token bubble.

This ignores the current TotalDamage boxes, since I wasnt sure what the Impale was for and didnt want to fiddle with that. But if TotalWounds changes, it does update.